Q2-Assignments/othoba-mobile-automation/requirements.txt
appium-python-client==2.0.0
selenium==4.0.0
pytest==7.0.0
pytest-html==3.1.1
